requirements.txt
numpy==1.24.3
pandas==2.0.2
Pillow==9.5.0
pytorch_lightning==2.0.2
PyYAML==6.0
scikit_learn==1.2.2
torch==2.0.1
torchmetrics==0.11.4
torchvision==0.15.2
